requirements.txt
django-admin-tools
south
mysql-python
django_compressor
BeautifulSoup
